+++ gulp/tasks/watch.js
@@ -1,0 +1,15 @@
+'use strict';
+
+/* Notes:
+   - gulp/tasks/scripts.js handles js recompiling with watchify
+   - gulp/tasks/browserSync.js watches and reloads compiled files
+*/
+
+var gulp = require( 'gulp' );
+var config = require( '../config' );
+
+gulp.task( 'watch', [ 'watchify', 'browserSync' ], function() {
+  gulp.watch( config.styles.cwd + '/**/*.less', [ 'styles' ] );
+  gulp.watch( config.images.src, [ 'images' ] );
+  gulp.watch( config.copy.files.src, [ 'copy:files' ] );
+} );